Remove deprecated message for PHP 7.4+ (#2015)
PHPMailer used in Joomla! 3 can't be updated to 6.x until Joomla! 4.0 is released because of minimum PHP version differences between Joomla! and PHPMailer. Could you please kindly accept this PR and release another version of PHPMailer 5.x that would not create a deprecated warning in PHP 7.4? This PR fixes the warning. Thanks :)